Renamed payment to webpayment screen, updated navigation
